pyproject.toml
[project]
name = "graph"
version = "3.0.11"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka[protobuf,schemaregistry]==2.12.0",
    "fastapi==0.119.0",
    "grpcio==1.75.1",
    "icecream==2.1.8",
    "protobuf==6.33.0",
    "pydantic==2.12.2",
    "pydantic-settings==2.11.0",
    "python-arango==8.2.2",
    "python-dateutil==2.9.0.post0",
    "python-keycloak==5.8.1",
    "pytz==2025.2",
    "uvicorn[standard]==0.37.0",
]

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
]
lint = [
    "ruff==0.14.0",
]
tests = [
    "httpx==0.28.1",
    "pytest==8.4.2",
    "pytest-mock==3.15.1",
    "testcontainers==4.13.2",
]
security = [
    "pip-audit==2.9.0",
]
migrations = [
]
